changed entrypoint script not to copy istgt.conf if it is already present (#26)

+++ entrypoint-istgtimage.sh
@@ -10,13 +10,15 @@ echo  "exit code:" $?
 echo "reference: "  $0 
 }
 
-cp /usr/local/etc/bkpistgt/istgt.conf /usr/local/etc/istgt/
+if [ ! -f "/usr/local/etc/istgt/istgt.conf" ];then
+	cp /usr/local/etc/bkpistgt/istgt.conf /usr/local/etc/istgt/
+	sed -i -n '/LogicalUnit section/,$!p' /usr/local/etc/istgt/istgt.conf
+fi
 cp /usr/local/etc/bkpistgt/istgtcontrol.conf /usr/local/etc/istgt/
 touch /usr/local/etc/istgt/auth.conf
 touch /usr/local/etc/istgt/logfile
 export externalIP=0.0.0.0
 service rsyslog start
-sed -i -n '/LogicalUnit section/,$!p' /usr/local/etc/istgt/istgt.conf
 
 exec /usr/local/bin/istgt &
